change the crud parameters for account, and the id(not long)

+++ src/main/java/nl/inholland/bank/repositories/AccountRepository.java
@@ -2,8 +2,6 @@ package nl.inholland.bank.repositories;
 
 import nl.inholland.bank.models.Account;
 import nl.inholland.bank.models.User;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,7 +9,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AccountRepository extends CrudRepository<Account, Long> {
+public interface AccountRepository extends CrudRepository<Account, Integer> {
    // get all the accounts according to a user id
     List<Account> findAllByUser(User user);
 }
